Add a test case about overriding the prelude
Probably not unexpected but without further protection against the
`prelude` command, one can easily prove unsound things, as this test
demonstrates.

It seems that either `lean4checker` should somehow ensure that only an
official prelude is used, or else the kernel (which implements `Nat.add`
differently) should check that the definition is indeed as expected.

I had to restructure the tests into their own library with a `glob`, as
_importing_ this file from another file that also imports the normal
prelude seems to fail with
```
./././Lean4CheckerTests.lean:1:0: error: import Lean4CheckerTests.OverridenPrelude failed, environment already contains 'Eq.refl' from Init.Prelude
```